Add base files for simulation project
This commit includes the following implementations along with limitations:

Simulation is limited to only one elevator at a time. Multiple elevator simulation can be achieved by using requests() and we will explore that option in the future.

Current implementation does not support changing arrival rates, elevator speed, weight/capacity limits, or specific probabilities for generating current and destination floors for different types of Person objects.

Statistical data visualisation is currently limited and does not include the plotting of histograms.

There is no GUI available for visualising the simulation. Currently, the output is displayed in the command line via print statements.

We will continue to explore options for improving these limitations in future commits.