Improve DES by pre-generating Person objects
Previously, the simulation of the entire building will timeout until a person is generated. This is entirely not accurate to how a DES should be.
`Simpy ` is not able to parallelise operations, to counter this limitation, this commit will generate a list of Person objects separately. This is done using the new `PersonList` class. The building then handles the `PersonList` during initialisation by placing the Person objects into their respective `Floor`. During the simulation, the `Floor` will update the status repeatedly in each iteration to check whether `Person` object can activate a call on the lifts based on the arrival time attribute of the `Person` via a simple comparison with the simulation time. If the condition is satisfied, the call to the lifts will be activated and when the allocation of lifts method is called it can add those floors to its path.

Additionally, the lift capacity should be taken into account in this commit.

Note that in this commit, most of the Python documentation are not updated or present. Please report any bugs in the issues.